BackEnd/requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
diff-match-patch==20241021
Django==5.2.1
django-anymail==13.0
django-ckeditor==6.7.0
django-ckeditor-5==0.2.17
django-cors-headers==4.7.0
django-dotenv==1.4.2
django-environ==0.12.0
django-import-export==4.3.7
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
humanize==4.12.3
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
shortuuid==1.0.13
sqlparse==0.4.4
tablib==3.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.4.0


# video 13